Elaborating on Background Threads (BgT). Now seperating every BgT into there own java class. adding new class for changing the enemy direction.
Trying to display a heavy message in loosing or winning, so two new TextViews are added, but still not activated.
MainActivity is extracting the two new textViews parsing them to GameController!

GameController is handling the BgT y putting them into own variables, so thy cna be cancelled if the game is Lost or Won.

Kind of 'deleting' the old AsyncMoveEnemies and 'adding' a new AsyncEnemyMove java class due to renaming!